hostinfo: refactor
